refactor(core): single rebindOrgScope helper for org-scoped context rebinding
resolve-org-from-path (path-scoped requests) and automation-context
(durable-run rebuilds) each hand-rolled their own subset of org-scoped
rebinding — which is exactly how ctx.orgFs stayed null on background
runs while objectStorage worked. One exported helper in context-factory
now rebinds thread storage, object storage, org-fs, and asset hoisters;
both callers use it, so the paths can't drift. New org-scoped fields go
in the helper, not at call sites.

automation-context's hand-constructed OrgScoped*Storage instances were
equivalent to setOrganizationId on the per-context wrappers (same
underlying SqlThreadStorage over the same db), so its threadStorage /
asyncResearchJobStorage deps are gone.
