+++ apps/mesh/src/api/app.ts
@@ -1411,8 +1411,6 @@ export async function createApp(options: CreateAppOptions = {}) {
 
   const automationContextFactory = createAutomationContextFactory({
     db: database.db,
-    threadStorage,
-    asyncResearchJobStorage,
   });
 
   // Stash deps for the DBOS workflow body. Safe to call before DBOS.launch():
